backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorgram.py==1.2.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
PyJWT==2.4.0
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
twilio==7.9.2
tzdata==2022.2
urllib3==1.26.9
